+++ Assets/deCrawl/Scrips/UI/CountDownTimer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c711e96b55b97345ae628679244904c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
